Clean up raster graphics package
