+++ src/app.js
@@ -1,104 +1,25 @@
 import './sass/editor.scss';
 import MapView from './views/map-view';
-import { detectRegion } from './graphics/raster/flood-fill';
-import PointIndex from './graphics/raster/point-index';
-import trackPath, { outline } from './graphics/raster/path-tracker';
+import MappingView from './views/mapping-view';
+import { Raster, Color } from './graphics/raster/raster';
+import { trackArea, trackOutline, trackPath } from './graphics/raster/tracking';
 import buildPolygon, { buildPolygonString } from './graphics/svg/polygon-builder';
 import { sendDistrict, fetchRegions } from './rest';
 // those imports will be replaced with configurable assets later on
 import atu from 'json!../examples/ukraine/regions.json';
 import mapImage from 'url?!../examples/ukraine/detailed-map.png';
 
-const createOptionElement = (value) => {
-  const option = document.createElement('option');
-  option.value = value;
-  option.text = value;
-  return option;
-};
-
-const removeAllChildElements = (element) => {
-  while (element.firstChild) {
-    element.removeChild(element.firstChild);
-  }
-};
-
-let path = [];
-let progressMap;
-
-class PropertiesView {
-  constructor(props) {
-    this.onPropertiesChange(props);
-    this.onStateChange({});
-  }
-
-  onPropertiesChange(props) {
-    this.props = props;
-  }
-
-  onStateChange(state) {
-    this.state = state;
-  }
-
-  updateState(newState) {
-    this.onStateChange(newState);
-    this.render();
-  }
-
-  updateProperties(props) {
-    this.onPropertiesChange(props);
-    this.render();
-  }
-
-  render() {
-    const regions = this.props.map(region => region.region).sort();
-    const selectedRegion = this.state.region || regions[0];
-    const districts = this.props.find(region => region.region === selectedRegion).districts.sort();
-
-    const $region = document.querySelector('.properties__region');
-    const $district = document.querySelector('.properties__district');
-    const $saveButton = document.querySelector('.properties__save-button');
-    const $regionOptions = regions.map(region => createOptionElement(region));
-    $regionOptions.forEach(option => $region.appendChild(option));
-    const $districtOptions = districts.map(district => createOptionElement(district));
-    $districtOptions.forEach(district => $district.appendChild(district));
-
-    $region.onchange = () => {
-      removeAllChildElements($district);
-      this.updateState({ region: $region.value });
-    };
-
-    $saveButton.onclick = () => {
-      if (!path.length) {
-        console.error('Please select district before saving');
-      }
-
-      const polygonString = buildPolygonString(path);
-      const region = $region.value;
-      const district = $district.value;
-      sendDistrict({
-        key: `${region}/${district}`.toLowerCase(),
-        region,
-        district,
-        polygon: polygonString,
-      }).then(() => {
-        progressMap.appendChild(buildPolygon(path));
-        path = []; // bad function. bad. side-effects. I will change it
-      });
-    };
-  }
-}
-
 function init() {
   const canvas = document.getElementById('map');
   const debugCanvas = document.querySelector('.debug__canvas');
-  progressMap = document.getElementById('progress');
+  const progressMap = document.getElementById('progress');
   progressMap.style.backgroundImage = `url(${mapImage})`;
 
   const mapView = new MapView(canvas, mapImage);
-  const propertiesView = new PropertiesView(atu);
+  const mappingView = new MappingView(atu);
 
   mapView.render();
-  propertiesView.render();
+  mappingView.render();
 
   fetchRegions().then(response => {
     response.json().then(mappedDistricts => {
@@ -115,8 +36,10 @@ function init() {
     const imageData = g.getImageData(0, 0, canvas.width, canvas.height);
     const x = evt.clientX - boundingRectangle.left;
     const y = evt.clientY - boundingRectangle.top;
-    const area = detectRegion(imageData, x, y, [255, 255, 255, 255]);
-    if (area.length === 0) {
+    const raster = new Raster(imageData);
+    const areaColor = new Color([255, 255, 255, 255]);
+    const area = trackArea(raster, x, y, areaColor);
+    if (area.isEmpty()) {
       return;
     }
     mapView.highlightRegion(area);
@@ -124,9 +47,23 @@ function init() {
     const districtCtx = debugCanvas.getContext('2d');
     districtCtx.fillStyle = '#cccccc';
     districtCtx.fillRect(0, 0, debugCanvas.width, debugCanvas.height);
-    const shapeOutline = outline(area);
-
-    path = trackPath(new PointIndex(shapeOutline));
+    const shapeOutline = trackOutline(area);
+    const path = trackPath(shapeOutline);
+    progressMap.appendChild(buildPolygon(path));
+    mappingView.onSave((region, district) => {
+      if (!path.length) {
+        console.error('Please select district before saving');
+      }
+      const polygonString = buildPolygonString(path);
+      sendDistrict({
+        key: `${region}/${district}`.toLowerCase(),
+        region,
+        district,
+        polygon: polygonString,
+      }).then(() => {
+        progressMap.appendChild(buildPolygon(path));
+      });
+    });
   };
 }
