--- src/debug.js
@@ -1,23 +1,0 @@
-import { setColor } from './graphics/raster/flood-fill';
-
-export function debugPaint(imageData, points, repaint, interval = 100, color = [0, 0, 255, 255]) {
-  let startTime = null;
-  let i = 0;
-
-  function render(timestamp) {
-    startTime = startTime || timestamp;
-    const progress = timestamp - startTime;
-    if (progress > interval) {
-      startTime = timestamp;  // resetting progress
-      const point = points[i];
-      setColor(imageData, point.x, point.y, color);
-      repaint();
-      i++;
-    }
-    if (i < points.length) {
-      requestAnimationFrame(render);
-    }
-  }
-
-  requestAnimationFrame(render);
-}
