+++ src/graphics/raster/point.js
@@ -1,0 +1,8 @@
+class Point {
+  constructor(x, y) {
+    this.x = x;
+    this.y = y;
+  }
+}
+
+export default Point;
