+++ src/graphics/svg/polygon-builder.js
@@ -1,10 +1,13 @@
+const listPointsAsString = (points) =>
+  points.listPoints()
+    .map(point => `${point.x},${point.y}`).join(' ');
+
 export default function buildPolygon(points) {
   const polygon = document.createElementNS('http://www.w3.org/2000/svg', 'polygon');
-  polygon.setAttribute('points', points.map(point => `${point.x},${point.y}`).join(' '));
+  polygon.setAttribute('points', listPointsAsString(points));
   return polygon;
 }
 
 export function buildPolygonString(points) {
-  const pointList = points.map(point => `${point.x},${point.y}`).join(' ');
-  return `<polygon points="${pointList}"/>`;
+  return `<polygon points="${listPointsAsString(points)}"/>`;
 }
