+++ test/unit/path-tracker-spec.js
@@ -1,13 +1,8 @@
 import { expect } from 'chai';
-import PointIndex from '../../src/graphics/raster/point-index';
-import trackPath from '../../src/graphics/raster/path-tracker';
+import PointIndex from '../../src/graphics/raster/points';
+import { trackPath } from '../../src/graphics/raster/tracking';
 
 describe('path tracker', () => {
-  /**
-   00xx00
-   0x00x0
-   00xx00
-   */
   it('should track path shape', () => {
     const outline = new PointIndex([
       { x: 2, y: 0 }, { x: 3, y: 0 },
@@ -16,7 +11,7 @@ describe('path tracker', () => {
 
     const path = trackPath(outline, { x: 1, y: 1 });
 
-    expect(path).to.eql([
+    expect(path.listPoints()).to.eql([
       { x: 1, y: 1 }, { x: 2, y: 0 }, { x: 3, y: 0 },
       { x: 4, y: 1 }, { x: 3, y: 2 }, { x: 2, y: 2 },
     ]);
@@ -32,7 +27,7 @@ describe('path tracker', () => {
 
     const path = trackPath(outline, { x: 0, y: 2 });
 
-    expect(path).to.eql([
+    expect(path.listPoints()).to.eql([
       { x: 0, y: 2 }, { x: 1, y: 1 }, { x: 1, y: 0 },
       { x: 2, y: 2 }, { x: 1, y: 3 },
     ]);
